pyproject.toml
[project]
name = "sterility-ai"
version = "0.1.0"
description = "AI-powered sterility detection for medical instruments"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow==11.3.0",
    "cleanvision==0.3.1",
    "ipykernel==6.30.1",
    "numpy==1.26.4",
    "pandas==2.3.2",
    "matplotlib==3.10.6",
    "albumentations==2.0.8",
    "opencv-python==4.11.0.86",
    "scikit-image==0.25.2",
    "pip==25.2",
    "streamlit==1.50.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]

[dependency-groups]
dev = [
    "black==25.9.0",
    "ruff==0.13.2",
]
